refactor(rsql): split rules validation helpers
